Refactoring component folder
