+++ project/src/components/Common/ExploreButton/index.tsx
@@ -2,7 +2,7 @@ import { fadeInThenBounceConfig } from '@/lib/animations/simple';
 
 import { FaAngleDoubleDown } from 'react-icons/fa';
 
-import Link from '@components/Link';
+import Link from '@/components/Common/Link';
 
 type ExploreButtonProps = {
     href: string;
