+++ project/src/components/Layout/Header/index.astro
@@ -1,5 +1,5 @@
 ---
-import Header from '@/components/Header';
+import Header from '@/components/Layout/Header';
 
 const pathname = new URL(Astro.request.url).pathname;
 ---
