+++ project/src/components/Pages/Home/Sections/Stats/LanguageSection.tsx
@@ -1,6 +1,7 @@
 import { motion } from 'motion/react';
 
 import { fadeInUpConfig } from '@/lib/animations/simple';
+import ProgressBar from '@/components/Common/ProgressBar';
 
 type LanguageSectionProps = {
     languages: Record<string, number>;
@@ -22,31 +23,9 @@ export default function LanguageSection({ languages }: LanguageSectionProps) {
                 {Object.entries(languages)
                     .sort(([, a], [, b]) => b - a)
                     .map(([language, percentage]) => (
-                        <LanguageBar key={language} language={language} percentage={percentage} />
+                        <ProgressBar key={language} skill={language} percentage={percentage} />
                     ))}
             </motion.div>
         </motion.div>
-    );
-}
-
-function LanguageBar({ language, percentage }: { language: string; percentage: number }) {
-    return (
-        <div className='grid grid-cols-[0.72fr_1fr_0.1fr] items-center rounded-lg border backdrop-blur-sm border-white/10 bg-white/[0.03] p-4 transition-all duration-200 hover:-translate-y-1 hover:border-white/20 hover:bg-white/[0.05] sm:grid-cols-[0.6fr_1fr_0.1fr] md:grid-cols-[0.2fr_1fr_0.1fr] lg:grid-cols-[0.3fr_1fr_0.1fr] xl:grid-cols-[0.2fr_1fr_0.1fr]'>
-            <span className='text-sm font-medium text-text-secondary'>{language}</span>
-            <div className='h-2 w-full rounded-full bg-white/[0.03]'>
-                <motion.div
-                    {...fadeInUpConfig}
-                    transition={{ ...fadeInUpConfig.transition, delay: 0.25 }}
-                    className='h-full rounded-full bg-gradient-to-r from-purple-500/50 to-blue-500/50 backdrop-blur-sm'
-                    style={{ width: `${percentage}%` }}
-                />
-            </div>
-            <motion.span
-                {...fadeInUpConfig}
-                transition={{ ...fadeInUpConfig.transition, delay: 0.3 }}
-                className='w-16 text-center text-sm text-text-secondary'>
-                {percentage.toFixed(1)}%
-            </motion.span>
-        </div>
     );
 }
