+++ project/src/components/Pages/Home/Sections/Stats/ProfileCard.tsx
@@ -5,7 +5,7 @@ import { fadeInUpConfig } from '@/lib/animations/simple';
 import { FaStar } from 'react-icons/fa6';
 import { VscGithubAlt, VscRepo } from 'react-icons/vsc';
 
-import Link from '@/components/Link';
+import Link from '@/components/Common/Link';
 
 import type { GithubStats } from '@/types/githubStats';
