+++ project/src/layouts/global/RootLayout.astro
@@ -2,7 +2,7 @@
 import Head from '@layouts/global/Head.astro';
 import ParticleBackground from '@/components/ParticleBackground';
 
-import Header from '@/components/Header';
+import Header from '@/components/Layout/Header';
 
 import '@/global.css';
