+++ project/src/pages/404.astro
@@ -2,7 +2,7 @@
 export const prerender = true;
 
 import Layout from '@layouts/global/RootLayout.astro';
-import Container from '@/components/Container';
+import Container from '@/components/Layout/Container';
 ---
 
 <Layout title='404'>
