+++ project/src/pages/about/index.astro
@@ -2,7 +2,7 @@
 export const prerender = true;
 
 import Layout from '@layouts/global/RootLayout.astro';
-import Container from '@/components/Container';
+import Container from '@/components/Layout/Container';
 import Greeting from '@/components/Pages/About/Sections/Greeting';
 import Overview from '@/components/Pages/About/Sections/Overview';
 const jobStatus = import.meta.env.JOB_STATUS;
