+++ project/src/pages/index.astro
@@ -2,7 +2,7 @@
 export const prerender = true;
 
 import Layout from '@layouts/global/RootLayout.astro';
-import Container from '@/components/Container';
+import Container from '@/components/Layout/Container';
 
 import Hero from '@/components/Pages/Home/Sections/Hero';
 import Stats from '@/components/Pages/Home/Sections/Stats/index.astro';
